requirements-web.txt
flask>=3.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
rembg[cpu]>=2.0.0  # [cpu] includes onnxruntime for background removal
opencv-contrib-python>=4.8.0
timm>=0.9.0
ultralytics>=8.0.0
segment-anything>=1.0
# LaMa AI inpainting (Python 3.12+); fallback to OpenCV if not installed
simple-lama-inpainting-updated>=0.1.3
# Google Veo (Vertex AI) — optional; use when USE_VEO=true
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
